Move unfold into own module
